Address Donald comments
